[AIE] Harmonize analyzeBranch representation
PseudoLoopEnd is very similar to a regular conditional branch. We need two Cond
elements in order to reconstruct the instruction, one is the opcode, the other
is the condition register for JZ/JNZ and the last-bundle label for PseudoLoopEnd

insertBranch needs to generate unconditional branch for FBB
even after PseudoLoopEnd.